Fixed the "not a module" problem
We were compiling with the wrong ecmascript version. If you look in
`tsconfig.json` you can see the change I made there

We also had some syntax errors in `app.ts` which I fixed up
and we had a final thing that I really should have remembered about
classes.

Basically, classes are like the "idea" of a thing. So the `Game` class
is the "idea" of a game. It can't actually run, it's too abstract. You
have to create a new game from the Game. You can see that change in
`app.test.ts`

We can walk through it in more detail in person